Move XmlStyle class to styles subpackage
