+++ viewdsl/src/main/java/splitties/viewdsl/core/styles/XmlStyle.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package splitties.viewdsl.core
+package splitties.viewdsl.core.styles
 
 import android.support.annotation.AttrRes
 import android.view.View
